+++ R/MEDseq.R
@@ -14,7 +14,7 @@
 #' 
 #' \code{\link{MEDseq_stderr}} is provided for computing the standard errors of the coefficients for the covariates in the gating network.
 #' 
-#' A dedicated plotting function exists for visualising various aspects of the results, using new methods as well as some existing methods from the \pkg{TraMineR} package.
+#' A dedicated plotting function \code{\link{plot.MEDseq}} exists for visualising various aspects of the results, using new methods as well as some existing methods from the \pkg{TraMineR} package.
 #' 
 #' Finally, the package also contains two data sets: \code{\link{biofam}} and \code{\link{mvad}}.
 #' 
@@ -22,8 +22,8 @@
 #' \itemize{
 #' \item{Type: }{Package}
 #' \item{Package: }{MEDseq}
-#' \item{Date: }{2019-12-10 (this version), 2019-08-24 (original release)}
 #' \item{Version: }{1.1.0}
+#' \item{Date: }{2020-03-30 (this version), 2019-08-24 (original release)}
 #' \item{Licence: }{GPL (>=2)}
 #' }
 #'
@@ -37,7 +37,7 @@
 #' \strong{Maintainer}: Keefe Murphy - <\email{keefe.murphy@@ucd.ie}>
 #' @references Murphy, K., Murphy, T. B., Piccarreta, R., and Gormley, I. C. (2019). Clustering longitudinal life-course sequences using mixtures of exponential-distance models. \emph{To appear}. <\href{https://arxiv.org/abs/1908.07963}{arXiv:1908.07963}>.
 #' @examples
-#' \dontshow{library(TraMineR)}
+#' \dontshow{suppressMessages(require(TraMineR))}
 #' # Load the MVAD data
 #' data(mvad)
 #' mvad$Location <- factor(apply(mvad[,5:9], 1L, function(x)