+++ inst/NEWS.md
@@ -1,12 +1,21 @@
 __MEDseq: Mixtures of Exponential-Distance Models with Covariates__   
 ===================================================================
 
-### New Features
+## MEDseq v1.1.0 - (_3<sup>rd</sup> release [minor update]: 2020-03-30_)
+### New Features, Improvements, and Bug Fixes
+* Significant efficiency gains when ignoring duplicates in the presence of weights:  
+    * before, unique cases were defined as unique sequence/covariates/weight combinations,  
+    * now, cases with different weights that are otherwise duplicates are treated as duplicates.
 * `MEDseq_stderr` is provided for computing the standard errors of the  
   coefficients for the covariates in the gating network via either the  
   weighted likelihood bootstrap or jackknife methods.
 * Small robustifications in the presence of empty components.
+* Fixed `get_MEDseq_results` when `what="MAP"` and non-noise models are chosen.
 * Fixed bug related to the colours used in the vignette plots.
+* Odds ratios now returned (and printed) when calling `summary` on `x$gating`.
+* Cosmetic fixes to `plot.MEDseq` when `type="clusters"` for small sample sizes.
+* Other small cosmetic plotting  & reference-formatting changes.
+* Spell-checking of documentation and fixes to `donttest` examples.
 
 ## MEDseq v1.0.1 - (_2<sup>nd</sup> release [patch update]: 2019-12-10_)
 ### Bug Fixes & Miscellaneous Edits